Split albert_shark_test.py into albert.py and test_shark_importer.py
