+++ tank/tflite/albert.py
@@ -5,8 +5,6 @@ import pytest
 
 model_path = "https://tfhub.dev/tensorflow/lite-model/albert_lite_base/squadv1/1?lite-format=tflite"
 
-
-
 # Inputs modified to be useful albert inputs.
 def generate_inputs(input_details):
     for input in input_details:
@@ -19,24 +17,12 @@ def generate_inputs(input_details):
     args.append(np.zeros(shape=input_details[2]["shape"], dtype=input_details[2]["dtype"]))
     return args
 
-# A specific case can be run by commenting different cases. Runs all the test
-# across cpu, gpu and vulkan according to available drivers.
-pytest_param = pytest.mark.parametrize(
-    ('dynamic', 'device'),
-    [
-        pytest.param(False, 'cpu'),
-        # TODO: Language models are failing for dynamic case..
-        pytest.param(True, 'cpu', marks=pytest.mark.skip),
-    ])
-
-
-@pytest_param
-def test_albert(dynamic, device):
+if __name__ == '__main__':
     my_shark_importer = SharkImporter(model_path=model_path,
                                       model_type="tflite",
                                       model_source_hub="tfhub",
-                                      device=device,
-                                      dynamic=dynamic,
+                                      device="cpu",
+                                      dynamic=False,
                                       jit_trace=True
                                       )
     input_details, output_details = my_shark_importer.get_model_details()
@@ -45,6 +31,3 @@ def test_albert(dynamic, device):
     my_shark_importer.compile()
     shark_results = my_shark_importer.forward()
     # print(shark_results)
-
-if __name__ == '__main__':
-    test_albert(False, "cpu")
